Review comments
- move cfg impls into strategy.rs
- remove ToStringCore and FromStringCore
- fix 'cargo check --tests' from API crate
- Fix comment typo
- Remove empty pod mod query
