+++ tiledb/api/src/filter/ftype.rs
@@ -1,21 +1,13 @@
-use std::fmt::{Display, Formatter, Result as FmtResult};
-
 use thiserror::Error;
 use tiledb_common::filter::{
     ChecksumType, CompressionData, CompressionType, FilterData,
 };
 
-use crate::{FromStringCore, ToStringCore};
-
 #[derive(Clone, Debug, Error)]
 pub enum Error {
     #[error("Invalid discriminant for {}: {0}", std::any::type_name::<FilterType>())]
     InvalidDiscriminant(u64),
-    #[error("Internal error formatting {0}")]
-    InternalString(FilterType),
 }
-
-type Result<T> = std::result::Result<T, Error>;
 
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub enum FilterType {
@@ -37,53 +29,6 @@ pub enum FilterType {
     Xor,
     WebP,
     Delta,
-}
-
-impl ToStringCore for FilterType {
-    type Error = Error;
-
-    fn to_string_core(&self) -> Result<String> {
-        let mut c_str = std::ptr::null::<std::os::raw::c_char>();
-        let res = unsafe {
-            ffi::tiledb_filter_type_to_str((*self).into(), &mut c_str)
-        };
-        if res == ffi::TILEDB_OK {
-            let c_msg = unsafe { std::ffi::CStr::from_ptr(c_str) };
-            Ok(String::from(c_msg.to_string_lossy()))
-        } else {
-            Err(Error::InternalString(*self))
-        }
-    }
-}
-
-impl FromStringCore for FilterType {
-    fn from_string_core(s: &str) -> Option<Self> {
-        let c_ftype =
-            std::ffi::CString::new(s).expect("Error creating CString");
-        std::ffi::CString::new(s).expect("Error creating CString");
-        let mut c_ret: u32 = 0;
-        let res = unsafe {
-            ffi::tiledb_filter_type_from_str(
-                c_ftype.as_c_str().as_ptr(),
-                &mut c_ret,
-            )
-        };
-
-        if res == ffi::TILEDB_OK {
-            FilterType::try_from(c_ret).ok()
-        } else {
-            None
-        }
-    }
-}
-
-impl Display for FilterType {
-    fn fmt(&self, f: &mut Formatter) -> FmtResult {
-        match self.to_string_core() {
-            Ok(s) => write!(f, "{}", s),
-            Err(e) => write!(f, "<FilterType: {}>", e),
-        }
-    }
 }
 
 impl From<FilterType> for ffi::tiledb_filter_type_t {
@@ -242,25 +187,5 @@ impl From<&FilterData> for FilterType {
 impl From<FilterData> for FilterType {
     fn from(value: FilterData) -> Self {
         FilterType::from(&value)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn filter_type_roundtrips() {
-        for i in 0..256 {
-            let maybe_ftype = FilterType::try_from(i);
-            if maybe_ftype.is_ok() {
-                let ftype = maybe_ftype.unwrap();
-                let ftype_str =
-                    ftype.to_string_core().expect("Error creating string.");
-                let str_ftype = FilterType::from_string_core(&ftype_str)
-                    .expect("Error round tripping filter type string.");
-                assert_eq!(str_ftype, ftype);
-            }
-        }
     }
 }
