--- tiledb/common/src/query/mod.rs
@@ -1,1 +1,0 @@
-pub mod read;
