+++ tiledb/sys/src/lib.rs
@@ -10,7 +10,6 @@ mod capi_enum;
 mod config;
 mod constants;
 mod context;
-mod datatype;
 mod dimension;
 mod domain;
 mod encryption;
@@ -42,7 +41,6 @@ pub use capi_enum::*;
 pub use config::*;
 pub use constants::*;
 pub use context::*;
-pub use datatype::*;
 pub use dimension::*;
 pub use domain::*;
 pub use encryption::*;
